牡丹江财政发票升级

+++ Brown/Action/FireAction.cs
@@ -466,7 +466,7 @@ namespace Brown.Action
 		/// <param name="nums"></param>
 		/// <param name="handler"></param>
 		/// <returns></returns>
-		public static int FireSales_Misc(string ac001, string itemId, int nums, string handler)
+		public static int FireSales_Misc(string ac001, string itemId, decimal nums, string handler)
 		{
 			//逝者编号
 			OracleParameter op_ac001 = new OracleParameter("ic_ac001", OracleDbType.Varchar2, 10);
